improve component and modules naming
