--- src/app/5-modules/my-first-component/my-first-component.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-my-first-component',
-  templateUrl: './my-first-component.component.html',
-  styleUrl: './my-first-component.component.scss'
-})
-export class MyFirstComponentComponent {
-
-}
