+++ src/app/5-modules/my-first-component/my-first.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-my-first-component',
+  templateUrl: './my-first.component.html',
+  styleUrl: './my-first.component.scss'
+})
+export class MyFirstComponent {
+
+}
