--- src/app/5-modules/my-second-component/my-second-component.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-my-second-component',
-  templateUrl: './my-second-component.component.html',
-  styleUrl: './my-second-component.component.scss'
-})
-export class MySecondComponentComponent {
-
-}
