+++ src/app/5-modules/my-second-component/my-second.component.scss
