+++ src/app/5-modules/my-second-component/my-second.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-my-second-component',
+  templateUrl: './my-second.component.html',
+  styleUrl: './my-second.component.scss'
+})
+export class MySecondComponent {
+
+}
